moved the fab_get/put/put_data funcs from utils to command
